requirements/requirements.txt
psycopg2-binary == 2.9.5
requests == 2.25.1
python-dotenv == 0.21.0